Reuse hardhat config for init script
